Added User Management page. Split code for several files. Fixed some mistakes.
