+++ .golangci.yml
@@ -128,7 +128,7 @@ issues:
     - path: _test\.go
       linters:
         - gomnd
-    - path: internal/db/pgdb/pgdb.go
+    - path: internal/db/pgdb
       linters:
         - gosec
         - gocritic
