+++ web/templates/header.html
@@ -10,15 +10,19 @@
   <title>{{ .title }}</title>
   <!--  Load VueJS with defer option-->
   <!--  From External -->
+  <!--  <script defer src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/js/fontawesome.min.js"></script>-->
   <!--  <script defer src="https://unpkg.com/vue@3"></scriptdefer>-->
   <!--  <script defer src="https://unpkg.com/validator@latest/validator.min.js"></script>-->
+
   <!--  From Internal-->
+<!--  <script defer src="/static/js/fontawesome.min.js"></script>-->
   <script defer src="/static/js/vue.global.js"></script>
   <script defer src="/static/js/validator.min.js"></script>
   <script defer src="/static/js/app.js"></script>
 
   <!-- Bootstrap core CSS -->
   <link href="/static/css/bootstrap.css" rel="stylesheet" crossorigin="anonymous">
+  <link rel="stylesheet" href="/static/css/font-awesome-all.min.css">
   <!-- Favicons -->
 <!--  <link rel="apple-touch-icon" href="https://getbootstrap.su/docs/5.0/assets/img/favicons/apple-touch-icon.png" sizes="180x180">-->
 <!--  <link rel="icon" href="https://getbootstrap.su/docs/5.0/assets/img/favicons/favicon-32x32.png" sizes="32x32" type="image/png">-->
