+++ web/templates/login.html
@@ -2,32 +2,48 @@
 <html lang="en">
 {{ template "header" .}}
 <body class="text-center">
-<p>
-    User Session is {{ .user_session }}
-</p>
+<!--<p>-->
+<!--    User Session is {{ .user_session }}-->
+<!--</p>-->
 <main class="form-signin">
-    <form action="/login" method="post">
-        <a href="/">
-            <img class="mb-4" src="/static/img/logo.png" alt="" width="330px">
-        </a>
-        <h1 class="h3 mb-3 fw-normal">Пожалуйста, войдите</h1>
+    {{ if .do_logout }}
+        <form action="/logout" method="get">
+            <a href="/">
+                <img class="mb-4" src="/static/img/logo.png" alt="" width="330px">
+            </a>
+            <p class="text-danger">
+                {{ .error_message }}
+            </p>
+            <button class="w-100 btn btn-lg btn-secondary" type="submit">Sign out</button>
+        </form>
+    {{ else }}
+        <form action="/login" method="post">
+            <a href="/">
+                <img class="mb-4" src="/static/img/logo.png" alt="" width="330px">
+            </a>
 
-        <div class="form-floating">
-            <input name="username" type="text" class="form-control" id="inputUsername" placeholder="Username">
-            <label for="inputUsername">Username</label>
-        </div>
-        <div class="form-floating">
-            <input name="password" type="password" class="form-control" id="inputPassword" placeholder="Password">
-            <label for="inputPassword">Password</label>
-        </div>
+            <p class="text-danger">
+                {{ .error_message }}
+            </p>
 
-        <div class="checkbox mb-3">
-            <label>
-                <input type="checkbox" value="remember-me"> Remember me
-            </label>
-        </div>
-        <button class="w-100 btn btn-lg btn-secondary" type="submit">Sign in</button>
-    </form>
+            <h1 class="h3 mb-3 text-secondary">Пожалуйста, войдите</h1>
+            <div class="form-floating">
+                <input name="username" type="text" class="form-control" id="inputUsername" placeholder="Username">
+                <label for="inputUsername">Username</label>
+            </div>
+            <div class="form-floating">
+                <input name="password" type="password" class="form-control" id="inputPassword" placeholder="Password">
+                <label for="inputPassword">Password</label>
+            </div>
+
+            <div class="checkbox mb-3">
+                <label>
+                    <input type="checkbox" value="remember-me"> Remember me
+                </label>
+            </div>
+            <button class="w-100 btn btn-lg btn-secondary" type="submit">Sign in</button>
+        </form>
+    {{ end }}
 </main>
 {{ template "footer" .}}
 </body>
