+++ web/templates/users.html
@@ -1,0 +1,68 @@
+{{define "users"}}
+<html lang="en">
+{{ template "header" .}}
+<body class="vueapp">
+{{ template "nav" .}}
+<main class="container">
+    <div class=" mb-4">
+        <div class="px-3 row">
+            <div class="border border-secondary col-1 link-caption">
+                ID
+            </div>
+            <div class="border border-secondary col-2 link-caption">
+                Username
+            </div>
+            <div class="border border-secondary col-2 link-caption">
+                First Name
+            </div>
+            <div class="border border-secondary col-2 link-caption">
+                Last Name
+            </div>
+            <div class="border border-secondary col-2 link-caption">
+                Email
+            </div>
+            <div class="border border-secondary col-2 link-caption">
+                Photo
+            </div>
+            <div class="border border-secondary col-1 link-caption">
+                Is Active?
+            </div>
+        </div>
+        <template v-for="user in users">
+            <div class="px-3 row">
+                <div class="border border-secondary col-1">
+                    {% user.id %}
+                </div>
+                <div class="border border-secondary col-2">
+                    {% user.username %}
+                </div>
+                <div class="border border-secondary col-2">
+                    {% user.first_name %}
+                </div>
+                <div class="border border-secondary col-2">
+                    {% user.last_name %}
+                </div>
+                <div class="border border-secondary col-2">
+                    {% user.email %}
+                </div>
+                <div class="border border-secondary col-2">
+                    {% user.phone %}
+                </div>
+                <div v-if="user.user_status" class="border border-secondary col-1">
+                    <a href="#"><img src="/static/img/user-enabled.png" width="16" height="16"></a>
+                    <a href="#"><img src="/static/img/edit_icon.svg" width="16" height="16"></a>
+                    <a href="#"><img src="/static/img/delete_icon.svg" width="16" height="16"></a>
+                </div>
+                <div v-else class="border border-secondary col-1">
+                    <a href="#"><img src="/static/img/user-disabled.png" width="16" height="16"></a>
+                    <a href="#"><img src="/static/img/edit_icon.svg" width="16" height="16"></a>
+                    <a href="#"><img src="/static/img/delete_icon.svg" width="16" height="16"></a>
+                </div>
+            </div>
+        </template>
+    </div>
+</main>
+{{ template "footer" .}}
+</body>
+</html>
+{{end}}
